IR: Separate internal nodes into separate source file
